RPC Server: Handle channel closedown when booting up.
Suppose the connection manager got a connection. Then the connection is closed down
right after. A server is now spawned. It gets the connection and tries to run on the
connection. But as soon as it calls into AMQP it will get a `closing` message back
since the connection will be closing down.

The patch here wraps the connection code such that we can throw connection errors back
if they happen. This makes the server try again in a while. We guess that the supervisor
tree will be terminated anyhow since the AMQP server is going down and then the
connection manager eventually will do the same.