Ignore failures when terminating.
We may be in a situation where we got a channel, but that channel died and
we are in the middle of a race to close down. In this case, the problem is
that the terminate call fails to operate properly. This change ignores
any error occurring from closing down.