Monitor callers and expire dead callers.
A problem in our current setup, which this commit will fix, is that we may have
a caller expire due to a timeout. Since we did not track this happening, we would
not remove the link from the caller to the internal "continuations" table. We place
a monitor on the caller and then we keep track of monitored processes. This means
that if a process dies and ends its lifetime, we can clean up our internal table correctly.

While here, improve status formatting, and also make sure that stray messages are ignored.