Trac #1081 - Make the supervisor tree robust against errors.
All supervisors were declared as being transient. This is wrong because if they
reach max_restart_intensity they send out a 'shutdown' which then means that the
tree will stop being there on the supervisor one level up. This is not what we want.
Rather, we want the tree's error to propagate, so enforce a restart by changing the
code to reflect it.