Trac #1088 - Split the configuration of AMQP down to the code.
We discussed this. Static configuration should go into the code and things you want to change
should go into the configuration proper. This means we expose new calls that can produce client
specifications for servers and clients directly. We also propose a way to configure the AMQP system
by providing a direct configuration file parser for this kind of proplist.